Initial Commit
This object provides a mechanism to override system and runtime APIs for
use with unit-testing.  This is a valuable ability when a test could not
normally be performed otherwise.

Examples:
Tests that relate to time
Tests that open files
Tests that require network communications
Tests that require system permissions not accessible to the test
machine.

However, there is no reason this class could not be used to hook API
calls for other purposes.