Added tests to verify send and recv
Added basic tests for UDP and TCP send and recv calls.
Added a mechanism to explicitly prohibit blocking calls, even for
blocking sockets.
Since this is for unit-testing, processing will not be held up any
longer than it takes to process a test.
Blocking is prohibited by default.
Blocking can be explicitly re-enabled.
If a call would block, for a blocking socket, the WOULD BLOCK error is
returned.
Simply setup the test to have the desired amount of data available to
avoid this faux pas